finish modular actions
