+++ client/src/reducers.js
@@ -1,4 +1,4 @@
-import * as Actions from './actions';
+import Actions from "./actions";
 
 const initialState = {
   pouches: [],
